Remove redundant storage of enclave DSU KeySSI
This commit removes unnecessary method that stored the enclave DSU KeySSI in environment variable. As per the previous implementation, the KeySSI was fetched and then stored in the environment variable which was redundant. Now, it just retrieves the KeySSI and proceeds further eliminating any scope of failures associated with unnecessary storing operation.