Added apiKey apis in opendsu
This commit includes the deletion of APIKeysClient.js and AdminAPIKeysTest.js files, as well as a comprehensive refactor of the code related to the API keys inside the 'opendsu-sdk' module. Additionally, the location of these files has been moved to increase modularity. Methods, including ones for becoming a system admin and associating an API key have been updated, leading to overall improved functionality.