Update secret handling in http module
Refactored the http module to better handle SSO secrets encryption keys, by removing the redundancy of assigning this key whenever it's not present in the headers. Improved the error management by differentiating between the absence of SSO key in the environment variables and its absence in the headers, enhancing the reliability of the system.