Simplify VersionlessDSU tests
This update significantly reduces the complexity of the tests in the resolver module of the opendsu-sdk. The excessive details have been pruned to focus on the core functionalities. The "mount" and "unmount" functionalities along with folder and file deletion tests have been removed from 'versionlessDSU' test files. This will allow for more straightforward testing and easier maintenance going forward.