Add option to save key SSI mappings in SDK
The changes introduce a new parameter to allow saving key SSI mappings to storage in the SDK. This provides flexibility in key management when initializing the Seed SSI Mapping and the Light DB enclave. Adjustments were also made to the related tests and the `double-check.json` file to accommodate these changes.