Added parse function to enclave instance
The parsing of keySSI identifiers now happens inside the storageStrategy instances across all modules. In each instance where keySSI was parsed, it has been replaced by storageStrategy.parse(), reducing dependency on the global keySSI space and allowing for more flexibility and modularity in managing storage strategies. A change has also been made to receive the enclave as a parameter during the creation in the keySSIFactory.