Add data object to CreateIdentity and improve data handling
The code changes allow the createIdentity function to accept a dataObject. This extra parameter is utilized to improve how data and the associated entity is handled. The updates also include error reporting in case of failure for data-related actions such as writing or commiting batches, and the ability to directly retrieve the data object from a resolvedDIDDocument.